H2O: Set the SO_REUSEPORT socket option on listening sockets
SO_REUSEPORT improves load balancing, especially when the
application server has a large number of CPU cores and/or
there are a lot of incoming connections.
